kotemap_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
geoalchemy2==0.15.2
alembic==1.13.3
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
redis==5.1.1
httpx==0.27.2
anthropic==0.34.2
firebase-admin==6.5.0
python-dotenv==1.0.1
psycopg[binary]==3.2.4
